Extract printing to syscall handler
