+++ crates/cheatnet/src/execution/cheatable_syscall_handler.rs
@@ -1,4 +1,4 @@
-use crate::execution::contract_print::{contract_print, PrintingResult};
+use crate::execution::contract_syscall_handler::ContractSyscallHandler;
 use crate::execution::{cheated_syscalls, syscall_hooks};
 use crate::state::CheatnetState;
 use anyhow::Result;
@@ -44,7 +44,7 @@ pub struct CheatableSyscallHandler<'a> {
 }
 
 impl<'a> CheatableSyscallHandler<'a> {
-    pub fn new(
+    pub fn wrap(
         syscall_handler: SyscallHintProcessor<'a>,
         cheatnet_state: &'a mut CheatnetState,
     ) -> Self {
@@ -93,13 +93,12 @@ impl HintProcessorLogic for CheatableSyscallHandler<'_> {
             }
         }
 
-        match contract_print(vm, maybe_extended_hint) {
-            PrintingResult::Printed => return Ok(()),
-            PrintingResult::Passed => (),
-            PrintingResult::Err(err) => return Err(err),
-        }
-        self.syscall_handler
-            .execute_hint(vm, exec_scopes, hint_data, constants)
+        ContractSyscallHandler::wrap(&mut self.syscall_handler).execute_hint(
+            vm,
+            exec_scopes,
+            hint_data,
+            constants,
+        )
     }
 
     /// Trait function to store hint in the hint processor by string.
