+++ crates/cheatnet/src/execution/mod.rs
@@ -3,7 +3,7 @@ pub mod cairo1_execution;
 pub mod calls;
 pub mod cheatable_syscall_handler;
 pub mod cheated_syscalls;
-pub mod contract_print;
+pub(crate) mod contract_syscall_handler;
 pub mod deprecated;
 pub mod entry_point;
 #[allow(clippy::module_name_repetitions)]
